chore: use svelte-kit package
