+++ packages/ssr/package.json
@@ -1,35 +1,65 @@
 {
-  "name": "@svebcomponents/ssr",
-  "version": "0.0.1",
-  "type": "module",
-  "scripts": {
-    "build": "tsc"
-  },
-  "exports": {
-    ".": {
-      "import": "./dist/src/index.js"
-    },
-    "./registry": {
-      "import": "./dist/src/renderIndex.js"
-    },
-    "./vite-plugin": {
-      "import": "./dist/src/vite-plugin-svebcomponent-ssr.js"
-    }
-  },
-  "devDependencies": {
-    "@lit-labs/ssr": "^3.3.1",
-    "@types/node": "22.13.10",
-    "eslint": "^9.18.0",
-    "eslint-config-prettier": "^10.0.1",
-    "eslint-plugin-svelte": "^2.46.1",
-    "estree-walker": "3.0.3",
-    "magic-string": "0.30.17",
-    "svelte": "^5.9.0",
-    "typescript": "^5.8.2",
-    "typescript-eslint": "^8.20.0",
-    "vite": "6.2.3"
-  },
-  "dependencies": {
-    "@lit-labs/ssr-dom-shim": "1.3.0"
-  }
+	"name": "@svebcomponents/ssr",
+	"version": "0.0.1",
+	"scripts": {
+		"dev": "vite dev",
+		"build": "vite build && pnpm prepack",
+		"preview": "vite preview",
+		"prepare": "svelte-kit sync || echo ''",
+		"prepack": "svelte-kit sync && svelte-package && publint",
+		"check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
+		"check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch",
+		"format": "prettier --write .",
+		"lint": "prettier --check . && eslint .",
+		"test:unit": "vitest",
+		"test:e2e": "playwright test"
+	},
+	"files": [
+		"dist",
+		"!dist/**/*.test.*",
+		"!dist/**/*.spec.*"
+	],
+	"type": "module",
+	"exports": {
+		".": {
+			"types": "./dist/index.d.ts",
+			"import": "./dist/index.js"
+		},
+		"./customElementWrapper": {
+			"svelte": "./dist/vitePlugin/CustomElementWrapper.svelte"
+		}
+	},
+	"peerDependencies": {
+		"svelte": "^5.0.0"
+	},
+	"devDependencies": {
+		"utils": "workspace:*",
+		"@eslint/compat": "^1.2.5",
+		"@eslint/js": "^9.18.0",
+		"@lit-labs/ssr": "^3.3.1",
+		"@lit-labs/ssr-dom-shim": "1.3.0",
+		"@playwright/test": "^1.49.1",
+		"@sveltejs/adapter-auto": "^4.0.0",
+		"@sveltejs/kit": "^2.16.0",
+		"@sveltejs/package": "^2.0.0",
+		"@sveltejs/vite-plugin-svelte": "^5.0.0",
+		"eslint": "^9.18.0",
+		"eslint-config-prettier": "^10.0.1",
+		"eslint-plugin-svelte": "^3.0.0",
+		"estree-walker": "3.0.3",
+		"globals": "^16.0.0",
+		"magic-string": "0.30.17",
+		"prettier": "^3.4.2",
+		"prettier-plugin-svelte": "^3.3.3",
+		"publint": "^0.3.2",
+		"svelte": "^5.0.0",
+		"svelte-check": "^4.0.0",
+		"typescript": "^5.0.0",
+		"typescript-eslint": "^8.20.0",
+		"vite": "^6.0.0",
+		"vitest": "^3.0.0"
+	},
+	"keywords": [
+		"svelte"
+	]
 }
