--- packages/ssr/src/index.ts
@@ -1,3 +1,0 @@
-import "./installShim.js";
-export * from "./render.js";
-export * from "./renderer.js";
