+++ packages/ssr/src/lib/index.ts
@@ -1,0 +1,4 @@
+import './elementRendererRegistry/installShim.js';
+export { ElementRendererRegistry } from './elementRendererRegistry/rendererRegistry.js';
+export { vitePluginSvebcomponentsSsr } from './vitePlugin/vite-plugin-svebcomponents-ssr.js';
+export { SvelteCustomElementRenderer } from './svelteCustomElementRenderer/svelteCustomElementRenderer.js';
