+++ packages/ssr/src/lib/vitePlugin/Client.svelte
