+++ packages/ssr/src/lib/vitePlugin/CustomElementWrapper.svelte
