--- packages/ssr/src/renderIndex.ts
@@ -1,5 +1,0 @@
-export * from "./rendererRegistry.js";
-export {
-  collectResult,
-  collectResultSync,
-} from "@lit-labs/ssr/lib/render-result.js";
