+++ packages/ssr/vite.config.ts
@@ -1,0 +1,6 @@
+import { defineConfig } from 'vite';
+import { sveltekit } from '@sveltejs/kit/vite';
+
+export default defineConfig({
+	plugins: [sveltekit()]
+});
